refactor: cli structure for help speed up, implement lazy loading
