+++ src/marbel/presets.py
@@ -1,12 +1,7 @@
-import pymc as pm
-import pandas as pd
 from importlib import resources
-from ete4 import Tree
 from enum import Enum
-import json
 
 __version__ = "0.1.0"
-
 
 CPMS_MEAN_LOG = 3.73411080985053
 CPMS_SD_LOG = 1.90214574373138
@@ -16,11 +11,12 @@ SPECIES_MEAN_LOG = 1.30069163364942
 SPECIES_SD_LOG = 1.33546646662937
 ORTHO_GROUPS_SHAPE = 0.675101585847872
 ORTHO_GROUPS_RATE = 0.00603147150265822
-
 DESEQ2_FITTED_A0 = 12.9105102
 DESEQ2_FITTED_A1 = 0.3325853
-
 DEFAULT_PHRED_QUALITY = 40
+DGE_LOG_2_CUTOFF_VALUE = 1
+MAX_SPECIES = 614
+MAX_ORTHO_GROUPS = 365813
 
 rank_distance = {
     "phylum": 0.6695424034631783,
@@ -31,9 +27,6 @@ rank_distance = {
 }
 
 
-MAX_SPECIES = 614
-MAX_ORTHO_GROUPS = 365813
-
 data_package = str(resources.files(__package__) / 'data')
 
 PATH_TO_GROUND_GENES = f"{data_package}/deduplicated_pangenome_EDGAR_Microbiome_JLAB2.fas.bgz"
@@ -42,18 +35,6 @@ PANGENOME_OVERVIEW = f"{data_package}/orthologues_processed_combined_all.parquet
 CDS_GENOMIC_LOCATIONS = f"{data_package}/cds_genomic_locations.parquet"
 SPECIES_PHYLO_TREE = f"{data_package}/EDGAR_all_species.newick"
 SPECIES_STATS = f"{data_package}/species_stats.json"
-
-pg_overview = pd.read_parquet(PANGENOME_OVERVIEW, engine='pyarrow')
-species_tree = Tree(open(SPECIES_PHYLO_TREE))
-DGE_LOG_2_CUTOFF_VALUE = 1
-species_stats_dict = json.load(open(SPECIES_STATS, 'r'))
-
-AVAILABLE_SPECIES = pg_overview.columns.to_list()[:MAX_SPECIES]
-
-with pm.Model() as model:
-    lognorm_dist = pm.Lognormal('read_counts', mu=READS_MEAN_LOG, sigma=READS_SD_LOG)  # reads
-    lognorm_dist2 = pm.Lognormal('species', mu=SPECIES_MEAN_LOG, sigma=SPECIES_SD_LOG)
-    orthologues = pm.Gamma('ortho', alpha=ORTHO_GROUPS_SHAPE, beta=ORTHO_GROUPS_RATE)  # ortho groups
 
 
 class ErrorModel(str, Enum):
